Moved Header, Block, Sequence, EmptySequence into maf.rb.
