+++ lib/bio/maf.rb
@@ -1,3 +1,4 @@
+require 'bio/maf/maf'
 require 'bio/maf/struct'
 require 'bio/maf/index'
 require 'bio/maf/parser'
